+++ lib/redis_failover/cli.rb
@@ -48,6 +48,21 @@ module RedisFailover
           options[:config_environment] = config_env
         end
 
+        opts.on('--node-strategy STRATEGY',
+         'Strategy used when determining availability of nodes (default: majority)') do |strategy|
+          options[:node_strategy] = strategy
+        end
+
+        opts.on('--failover-strategy STRATEGY',
+         'Strategy used when failing over to a new node (default: latency)') do |strategy|
+          options[:failover_strategy] = strategy
+        end
+
+        opts.on('--required-node-managers COUNT',
+         'Required Node Managers that must be reachable to determine node state (default: 1)') do |count|
+          options[:required_node_managers] = Integer(count)
+        end
+
         opts.on('-h', '--help', 'Display all options') do
           puts opts
           exit
@@ -59,7 +74,7 @@ module RedisFailover
         options = from_file(config_file, options[:config_environment])
       end
 
-      if required_options_missing?(options)
+      if invalid_options?(options)
         puts parser
         exit
       end
@@ -68,7 +83,7 @@ module RedisFailover
     end
 
     # @return [Boolean] true if required options missing, false otherwise
-    def self.required_options_missing?(options)
+    def self.invalid_options?(options)
       return true if options.empty?
       return true unless options.values_at(:nodes, :zkservers).all?
       false
@@ -111,6 +126,14 @@ module RedisFailover
       # assume password is same for all redis nodes
       if password = options[:password]
         options[:nodes].each { |opts| opts.update(:password => password) }
+      end
+
+      if node_strategy = options[:node_strategy]
+        options[:node_strategy] = node_strategy.to_sym
+      end
+
+      if failover_strategy = options[:failover_strategy]
+        options[:failover_strategy] = failover_strategy.to_sym
       end
 
       options