+++ spec/node_watcher_spec.rb
@@ -6,7 +6,7 @@ module RedisFailover
       @node_states = {}
     end
 
-    def notify_state(node, state)
+    def notify_state(node, state, latency = nil)
       @node_states[node] = state
     end
 
@@ -50,7 +50,7 @@ module RedisFailover
           watcher.watch
           sleep(3)
           watcher.shutdown
-          node_manager.state_for(node).should == :syncing
+          node_manager.state_for(node).should == :available
         end
       end
     end