+++ spec/support/node_manager_stub.rb
@@ -13,9 +13,11 @@ module RedisFailover
       [master, slave].each { |node| node.extend(RedisStubSupport) }
       master.make_master!
       slave.make_slave!(master)
+      @nodes = [master, slave]
       @unavailable = []
       @master = master
       @slaves = [slave]
+      @failover_strategy = Object.new
       @nodes_discovered = true
     end
 
@@ -33,34 +35,53 @@ module RedisFailover
     end
 
     def stop_processing
-      @queue << nil
       @thread.value
     end
 
     def force_unavailable(node)
       start_processing
       node.redis.make_unavailable!
-      notify_state(node, :unavailable)
+      snapshot = OpenStruct.new(
+        :node => node,
+        :available_count => 0,
+        :unavailable_count => 1,
+        :node_managers => ['nm'])
+      update_master_state(node, node => snapshot)
       stop_processing
     end
 
     def force_available(node)
       start_processing
       node.redis.make_available!
-      notify_state(node, :available)
+      snapshot = OpenStruct.new(
+        :node => node,
+        :available_count => 1,
+        :unavailable_count => 0,
+        :node_managers => ['nm'])
+      update_master_state(node, node => snapshot)
       stop_processing
     end
 
     def force_syncing(node, serve_stale_reads)
       start_processing
       node.redis.force_sync_with_master(serve_stale_reads)
-      notify_state(node, :syncing)
+      snapshot = OpenStruct.new(
+        :node => node,
+        :available_count => 1,
+        :unavailable_count => 0,
+        :node_managers => ['nm'])
+      update_master_state(node, node => snapshot)
       stop_processing
     end
 
-    def initialize_path; end
-    def delete_path; end
-    def create_path; end
-    def write_state; end
+    def failover_strategy_candidate(snapshots)
+      @slaves.pop
+    end
+
+    def delete_path(*args); end
+    def create_path(*args); end
+    def write_state(*args); end
+    def wait_until_master; end
+    def current_node_snapshots; {} end
   end
 end