+++ Changes.md
@@ -1,3 +1,20 @@
+HEAD
+-----------
+- redis_failover now supports distributed monitoring among the node managers! Previously, the node managers were only used
+as a means of redundancy in case a particular node manager crashed. Starting with version 1.0 of redis_failover, the node
+managers will all periodically report their health report/snapshots. The primary node manager will utilize a configurable
+"node strategy" to determine if a particular node is available or unavailable.
+- redis_failover now supports a configurable "failover strategy" that's consulted when performing a failover. Currently,
+a single strategy is provided that takes into account the average latency of the last health check to the redis server.
+
+0.9.7.2
+-----------
+- Add support for Redis#client's location method. Fixes a compatibility issue with redis_failover and Sidekiq.
+
+0.9.7.1
+-----------
+- Stop repeated attempts to acquire exclusive lock in Node Manager (#36)
+
 0.9.7
 -----------
 - Stubbed Client#client to return itself, fixes a fork reconnect bug with Resque (dbalatero)